+++ code/lab_series/lab3/lu_rows_blocks_master-slave_variant11/matrix_vector_utils.h
@@ -1,13 +1,13 @@
 struct Vector {
     int size;
-    int* data;
+    double* data;
 };
 
 struct Matrix
 {
     int rows;
     int cols;
-    int **data;
+    double** data;
 };
 
 struct Vector* read_vector(const char* filename);
